+++ orion-ops-ui/src/views/user/user/components/user-grant-roles-form-modal.vue
@@ -64,6 +64,7 @@
   const open = (record: any) => {
     renderForm(record);
     setVisible(true);
+    // 加载用户角色
     loadRoles();
   };